@@ -16,28 +16,29 @@
         }
     ],
     "require": {
-        "php": "^7.2 || ^8.0",
+        "php": "~8.1 || ~8.2",
         "doctrine/doctrine1": "^1.2.0",
-        "jdorn/sql-formatter": "^1.2.16",
-        "symfony/cache": "^4.3.3|^5.0",
-        "symfony/config": "^4.3.3|^5.0",
-        "symfony/console": "^3.4.30|^4.3.3|^5.0",
-        "symfony/dependency-injection": "^4.3.3|^5.0",
-        "symfony/framework-bundle": "^3.4.30|^4.3.3|^5.0",
-        "symfony/service-contracts": "^1.1.1|^2.0"
+        "doctrine/sql-formatter": "^1.0.1",
+        "symfony/cache": "^5.4 || ^6.0 || ^7.0",
+        "symfony/config": "^5.4 || ^6.0 || ^7.0",
+        "symfony/console": "^5.4 || ^6.0 || ^7.0",
+        "symfony/dependency-injection": "^5.4 || ^6.0 || ^7.0",
+        "symfony/framework-bundle": "^5.4 || ^6.0 || ^7.0",
+        "symfony/service-contracts": "^1.1.1 || ^2.0 || 3.0"
     },
     "require-dev": {
-        "diablomedia/php-cs-fixer-config": "^2.0.0",
-        "phpstan/phpstan": "0.12.88",
+        "diablomedia/php-cs-fixer-config": "^2.1.0",
+        "friendsofphp/php-cs-fixer": "^3.20.0",
+        "phpstan/phpstan": "1.10.60",
         "psalm/plugin-symfony": "^2.2.4",
         "symfony/property-info": "^4.3.3|^5.0",
-        "symfony/proxy-manager-bridge": "^3.4|^4.3.3|^5.0",
-        "symfony/twig-bridge": "^3.4.30|^4.3.3|^5.0",
-        "symfony/validator": "^3.4.30|^4.3.3|^5.0",
-        "symfony/web-profiler-bundle": "^3.4.30|^4.3.3|^5.0",
-        "symfony/yaml": "^3.4.30|^4.3.3|^5.0",
-        "twig/twig": "^1.34|^2.12|^3.0",
-        "vimeo/psalm": "4.7.3"
+        "symfony/proxy-manager-bridge": "^5.4 || ^6.0 || ^7.0",
+        "symfony/twig-bridge": "^5.4 || ^6.0 || ^7.0",
+        "symfony/validator": "^5.4 || ^6.0 || ^7.0",
+        "symfony/web-profiler-bundle": "^5.4 || ^6.0 || ^7.0",
+        "symfony/yaml": "^5.4 || ^6.0 || ^7.0",
+        "twig/twig": "^1.34 || ^2.12 || ^3.0",
+        "vimeo/psalm": "4.30.0"
     },
     "config": {
         "sort-packages": true
